agentic-repos-program/4-chat-based AI assistant/1_sync_agent/pyproject.toml
[project]
name = "sync-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai-agents>=0.0.16",
    "python-dotenv>=1.1.0",
]
